feat: optimized primality tests
First, capped the number of tests for the Miller-Rabin algorithm to 40
test (see the reason in the comments of `estimate_num_tests`).
Secondly, checked the divisibility of the prime candidates by a set of
small prime numbers (see the reason in the comments of
`preliminary_checks`). Those two optimizations (the latter more than the
former) makes the prime generation algorithm very fast.